Clean up refresh intervals to use seconds everywhere
BREAKING CHANGE: Changed from minutes/negative-value-hacks to clean seconds-only

This is a complete refactor removing all the ugly workarounds:
- Removed negative value encoding (-6=5s, -5=10s, etc.)
- Changed config.maxsleep from int8_t to uint16_t (0-65535 seconds)
- Updated all proto.h files to document seconds-only usage
- Simplified all AP firmware code to use seconds directly
- Updated web interface dropdown to use clean values (5, 10, 15...3600)
- Fixed TLSR tag firmware to expect seconds
- Updated JavaScript timeout logic to use seconds

Benefits:
- Much cleaner code with no magic negative values
- Supports up to 18.2 hours (65535 seconds) vs old limit
- Easier to understand and maintain
- Default changed from confusing "0/10" to clean "40 seconds"

Web interface now shows:
- 5s, 10s, 15s, 20s, 30s, 40s (default)
- 5min (300s), 10min (600s), 30min (1800s), 1hr (3600s)

+++ ESP32_AP-Flasher/src/web.cpp
@@ -601,7 +601,7 @@ void init_web() {
             updateLanguageFromConfig();
         }
         if (request->hasParam("maxsleep", true)) {
-            config.maxsleep = static_cast<int8_t>(request->getParam("maxsleep", true)->value().toInt());
+            config.maxsleep = static_cast<uint16_t>(request->getParam("maxsleep", true)->value().toInt());
         }
         if (request->hasParam("stopsleep", true)) {
             config.stopsleep = static_cast<uint8_t>(request->getParam("stopsleep", true)->value().toInt());
